refactor: split testfactory/generate_txs.go in txs.go` and blob.go
